Create cardcreator.py
This is a card generation program that somewhat follows the ruleset for Savage Worlds Freedom Squadron.  I didn't want to buy the cards for our game night, so I downloaded images from google and wrote a program to create over 70 cards for our game deck.  I did not own the cards, nor do I own Freedom Squadron.  This was for private use.

For a simpler program, try the cardtest.py program I uploaded.  That one will create a specific card based on the filename you give it, and save it based on the filename and path you give it.